pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bridgeme"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "SSH relay-based reverse shell tool for IT troubleshooting"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "paramiko>=2.7.0",
    "click>=7.0",
    "psutil>=5.0.0",
    "colorama>=0.4.0",
]

[project.optional-dependencies]
windows = ["pywinpty>=1.1.0"]
dev = ["pytest>=6.0", "pytest-cov", "black", "flake8"]

[project.scripts]
bridgeme = "bridgeme.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/bridgeme"
"Bug Reports" = "https://github.com/yourusername/bridgeme/issues"
"Source" = "https://github.com/yourusername/bridgeme"

[tool.setuptools.packages.find]
where = ["."]
include = ["bridgeme*"]
exclude = ["tests*"]